Move to Best Practices, simplified test
