+++ docs/pages/release_notes.md
@@ -26,7 +26,7 @@ This is a {{ site.pmd.release_type }} release.
 
 ### 🌟️ New and Changed Rules
 #### New Rules
-*   The new java rule  {% rule java/errorprone/TypeNameMismatch %} finds types that are not defined in a .java file
+*   The new java rule  {% rule java/bestpractices/TypeNameMismatch %} finds types that are not defined in a .java file
     with the same name. Enforcing a match between source file name and type name makes it easier to 
     find source code for given type.
