+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/TypeNameMismatchRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.rule.codestyle;
+package net.sourceforge.pmd.lang.java.rule.bestpractices;
 
 import org.apache.commons.lang3.StringUtils;
