+++ pmd-java/src/main/resources/category/java/bestpractices.xml
@@ -1498,6 +1498,40 @@ class Foo{
         </example>
     </rule>
 
+    <rule name="TypeNameMismatch"
+          language="java"
+          since="7.28.0"
+          message="Top-level type {0} should be defined in a file called {0}.java"
+          class="net.sourceforge.pmd.lang.java.rule.bestpractices.TypeNameMismatchRule"
+          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_bestpractices.html#typenamemismatch">
+        <description>
+            <![CDATA[
+Each top-level type should be defined in its own file with a matching name.
+]]>
+        </description>
+        <priority>3</priority>
+        <example>
+            <![CDATA[
+// content of Foo.java
+
+class Bar { // violation, should be defined in Bar.java
+
+}
+]]>
+        </example>
+        <example>
+            <![CDATA[
+// content of Baz.java
+class Baz {
+  class Foo { } // OK, nested classes are allowed
+}
+
+class Bar { // violation, second top level class should be in its own file
+}
+]]>
+        </example>
+    </rule>
+
     <rule name="UnitTestAssertionsShouldIncludeMessage"
           language="java"
           since="1.04"
