+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/bestpractices/xml/TypeNameMismatch.xml
